Simple sql statement extracted
