+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-scaling/src/test/java/org/apache/shardingsphere/integration/data/pipeline/cases/command/CommonSQLCommand.java
@@ -51,12 +51,6 @@ public final class CommonSQLCommand {
     @XmlElement(name = "simple-insert-order")
     private String simpleInsertOrder;
     
-    @XmlElement(name = "update-order")
-    private String updateOrder;
-    
-    @XmlElement(name = "update-order-item")
-    private String updateOrderItem;
-    
     @XmlElement(name = "delete-order")
     private String deleteOrder;
